backend/requirements.txt
Flask==3.1.0
flask-cors==6.0.0
SQLAlchemy>=2.0.40
psycopg2-binary
alembic>=1.15.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
marshmallow>=3.20.1
pytest>=7.4.0
pytest-mock>=3.11.1
